Remove default case to fail fast. Move safe pattern logic to separate file
